fix vuln+weak interactions

+++ Status/Progression.cs
@@ -29,9 +29,9 @@ namespace StS
     public class AttackProgression
     {
         public string Desc { get; set; }
-        public Func<IEnumerable<int>, IEnumerable<int>> Fun { get; set; }
+        public Func<IEnumerable<double>, IEnumerable<double>> Fun { get; set; }
 
-        public AttackProgression(string desc, Func<IEnumerable<int>, IEnumerable<int>> func)
+        public AttackProgression(string desc, Func<IEnumerable<double>, IEnumerable<double>> func)
         {
             Desc = desc;
             Fun = func;
